requirements.txt
branca==0.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
folium==0.14.0
fonttools==4.60.1
geographiclib==2.1
geopy==2.3.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.7.2
numpy==1.24.4
Office365-REST-Python-Client==2.5.11
packaging==25.0
pandas==2.0.3
pillow==12.0.0
pyparsing==3.0.9
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6-WebEngine==6.5.0
PyQt6-WebEngine-Qt6==6.10.0
PyQt6_sip==13.10.2
pyqtgraph==0.13.3
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.0.4
requests==2.32.5
scipy==1.10.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
